holy shit

+++ code/modules/projectiles/projectile.dm
@@ -18,6 +18,9 @@
 	pass_flags = PASSTABLE
 	mouse_opacity = 0
 	embed_mult = 1
+
+	var/testing = FALSE
+
 	var/bumped = FALSE		//Prevents it from hitting more than one guy at once
 	var/hitsound_wall = "ricochet"
 	var/list/mob_hit_sound = list('sound/effects/gore/bullethit2.ogg', 'sound/effects/gore/bullethit3.ogg') //Sound it makes when it hits a mob. It's a list so you can put multiple hit sounds there.
@@ -107,7 +110,6 @@
 
 	. = ..()
 
-
 /obj/item/projectile/is_hot()
 	if (damage_types[BURN])
 		return damage_types[BURN] * heat
@@ -151,21 +153,29 @@
 		damage_types[damage_type] += newdamages[damage_type]
 
 /obj/item/projectile/proc/on_hit(atom/target, def_zone = null)
-	if(!isliving(target))	return 0
-	if(isanimal(target))	return 0
+	if(!isliving(target))
+		return FALSE
+	if(isanimal(target))
+		return FALSE
 	var/mob/living/L = target
-	L.apply_effects(stun, weaken, paralyze, irradiate, stutter, eyeblur, drowsy)
+	if (!testing)
+		L.apply_effects(stun, weaken, paralyze, irradiate, stutter, eyeblur, drowsy)
 	return TRUE
 
 // generate impact effect
-/obj/item/projectile/proc/on_impact(atom/A)
-	impact_effect(effect_transform)
-	if(luminosity_ttl && attached_effect)
-		spawn(luminosity_ttl)
-		qdel(attached_effect)
+/obj/item/projectile/proc/on_impact(atom/A, override = FALSE)
+	if (!testing)
+		if (!override)
+			impact_effect(effect_transform)
+			if(luminosity_ttl && attached_effect)
+				spawn(luminosity_ttl)
+				qdel(attached_effect)
 
-	if(!ismob(A))
-		playsound(src, hitsound_wall, 50, 1, -2)
+			if(!ismob(A))
+				playsound(src, hitsound_wall, 50, 1, -2)
+
+	else if (testing)
+		SEND_SIGNAL(src, COMSIG_TRACE_IMPACT, src, A)
 	return
 
 /obj/item/projectile/multiply_pierce_penetration(newmult)
@@ -209,7 +219,7 @@
 		qdel(src)
 		return FALSE
 
-	if(proj_sound)
+	if(proj_sound && (!testing))
 		playsound(proj_sound)
 
 	original = target
@@ -221,7 +231,7 @@
 
 	return FALSE
 
-//called to launch a projectile from a gun
+//called to launch a projectile from a gun, not called by testing projectiles
 /obj/item/projectile/proc/launch_from_gun(atom/target, mob/user, obj/item/gun/launcher, target_zone, x_offset=0, y_offset=0, angle_offset)
 	if(user == target) //Shooting yourself
 		user.bullet_act(src, target_zone)
@@ -611,50 +621,53 @@
 
 
 	if(result == PROJECTILE_FORCE_MISS)
-		if(!silenced)
-			visible_message(SPAN_NOTICE("\The [src] misses [target_mob] narrowly!"))
-		return FALSE
+		if (!testing) //doesnt matter, we collided with something, NIKO COME BACK HERE AND REVIEW THIS
+			if(!silenced)
+				visible_message(SPAN_NOTICE("\The [src] misses [target_mob] narrowly!"))
+			return FALSE
 
 	//hit messages
-	if(silenced)
-		to_chat(target_mob, SPAN_DANGER("You've been hit in the [parse_zone(def_zone)] by \the [src]!"))
-	else
-		visible_message(SPAN_DANGER("\The [target_mob] is hit by \the [src] in the [parse_zone(def_zone)]!"))//X has fired Y is now given by the guns so you cant tell who shot you if you could not see the shooter
-
-	playsound(target_mob, pick(mob_hit_sound), 40, 1)
-
-	//admin logs
-	if(!no_attack_log)
-		if(ismob(firer))
-
-			var/attacker_message = "shot with \a [src.type]"
-			var/victim_message = "shot with \a [src.type]"
-			var/admin_message = "shot (\a [src.type])"
-
-			admin_attack_log(firer, target_mob, attacker_message, victim_message, admin_message)
+	if (!testing)
+		if(silenced)
+			to_chat(target_mob, SPAN_DANGER("You've been hit in the [parse_zone(def_zone)] by \the [src]!"))
 		else
-			target_mob.attack_log += "\[[time_stamp()]\] <b>UNKNOWN SUBJECT (No longer exists)</b> shot <b>[target_mob]/[target_mob.ckey]</b> with <b>\a [src]</b>"
-			msg_admin_attack("UNKNOWN shot [target_mob] ([target_mob.ckey]) with \a [src] (<A HREF='?_src_=holder;adminplayerobservecoodjump=1;X=[target_mob.x];Y=[target_mob.y];Z=[target_mob.z]'>JMP</a>)")
+			visible_message(SPAN_DANGER("\The [target_mob] is hit by \the [src] in the [parse_zone(def_zone)]!"))//X has fired Y is now given by the guns so you cant tell who shot you if you could not see the shooter
+
+		playsound(target_mob, pick(mob_hit_sound), 40, 1)
+
+		//admin logs
+		if(!no_attack_log)
+			if(ismob(firer))
+
+				var/attacker_message = "shot with \a [src.type]"
+				var/victim_message = "shot with \a [src.type]"
+				var/admin_message = "shot (\a [src.type])"
+
+				admin_attack_log(firer, target_mob, attacker_message, victim_message, admin_message)
+			else
+				target_mob.attack_log += "\[[time_stamp()]\] <b>UNKNOWN SUBJECT (No longer exists)</b> shot <b>[target_mob]/[target_mob.ckey]</b> with <b>\a [src]</b>"
+				msg_admin_attack("UNKNOWN shot [target_mob] ([target_mob.ckey]) with \a [src] (<A HREF='?_src_=holder;adminplayerobservecoodjump=1;X=[target_mob.x];Y=[target_mob.y];Z=[target_mob.z]'>JMP</a>)")
 
 	//sometimes bullet_act() will want the projectile to continue flying
 	if (result == PROJECTILE_CONTINUE)
 		return FALSE
 
-	if(target_mob.mob_classification & CLASSIFICATION_ORGANIC)
-		var/turf/target_loca = get_turf(target_mob)
-		var/mob/living/L = target_mob
-		if(damage_types[BRUTE] > 10)
-			var/splatter_dir = dir
-			if(starting)
-				splatter_dir = get_dir(starting, target_loca)
-				target_loca = get_step(target_loca, splatter_dir)
-			var/blood_color = "#C80000"
-			if(ishuman(target_mob))
-				var/mob/living/carbon/human/H = target_mob
-				blood_color = H.form.blood_color
-			new /obj/effect/overlay/temp/dir_setting/bloodsplatter(target_mob.loc, splatter_dir, blood_color)
-			if(target_loca && prob(50))
-				target_loca.add_blood(L)
+	if (!testing)
+		if(target_mob.mob_classification & CLASSIFICATION_ORGANIC)
+			var/turf/target_loca = get_turf(target_mob)
+			var/mob/living/L = target_mob
+			if(damage_types[BRUTE] > 10)
+				var/splatter_dir = dir
+				if(starting)
+					splatter_dir = get_dir(starting, target_loca)
+					target_loca = get_step(target_loca, splatter_dir)
+				var/blood_color = "#C80000"
+				if(ishuman(target_mob))
+					var/mob/living/carbon/human/H = target_mob
+					blood_color = H.form.blood_color
+				new /obj/effect/overlay/temp/dir_setting/bloodsplatter(target_mob.loc, splatter_dir, blood_color)
+				if(target_loca && prob(50))
+					target_loca.add_blood(L)
 
 	return TRUE
 
@@ -683,28 +696,30 @@
 			return FALSE
 	if(iscarbon(A))
 		var/mob/living/carbon/C = A
-		var/obj/item/shield/S
-		for(S in get_both_hands(C))
+		if (!testing)
+			var/obj/item/shield/S
+			for(S in get_both_hands(C))
+				if(S && S.block_bullet(C, src, def_zone))
+					on_hit(S,def_zone)
+					qdel(src)
+					return TRUE
+				break //Prevents shield dual-wielding
+			S = C.get_equipped_item(slot_back)
 			if(S && S.block_bullet(C, src, def_zone))
 				on_hit(S,def_zone)
 				qdel(src)
 				return TRUE
-			break //Prevents shield dual-wielding
-		S = C.get_equipped_item(slot_back)
-		if(S && S.block_bullet(C, src, def_zone))
-			on_hit(S,def_zone)
-			qdel(src)
-			return TRUE
 
 	if(ismob(A))
 		var/mob/M = A
 		if(isliving(A))
 			//if they have a neck grab on someone, that person gets hit instead
-			var/obj/item/grab/G = locate() in M
-			if(G && G.state >= GRAB_NECK && G.affecting.health >= 0) //Cant use dead or dieing bodies as a shield as they are limp
-				visible_message(SPAN_DANGER("\The [M] uses [G.affecting] as a shield!"))
-				if(Bump(G.affecting, TRUE))
-					return //If Bump() returns 0 (keep going) then we continue on to attack M.
+			if (!testing)
+				var/obj/item/grab/G = locate() in M
+				if(G && G.state >= GRAB_NECK && G.affecting.health >= 0) //Cant use dead or dieing bodies as a shield as they are limp
+					visible_message(SPAN_DANGER("\The [M] uses [G.affecting] as a shield!"))
+					if(Bump(G.affecting, TRUE))
+						return //If Bump() returns 0 (keep going) then we continue on to attack M.
 			passthrough = !attack_mob(M, distance)
 		else
 			passthrough = FALSE //so ghosts don't stop bullets
@@ -787,7 +802,7 @@
 
 			//Clone dosnt get removed do to being rare same as o2
 
-			if(!hitscan)
+			if(!hitscan || testing) //TODO: test
 				speed_drop_off = range_shot / 500 //How far we were shot - are affective range. This one is for speed drop off
 				// Every 500 steps = 1 step delay , 50 is 0.1 (thats huge!), 5 0.01
 				step_delay = step_delay + speed_drop_off
@@ -810,12 +825,13 @@
 					if(Bump(original))
 						return
 
-		if(first_step)
-			muzzle_effect(effect_transform)
-			first_step = FALSE
-		else if(!bumped)
-			tracer_effect(effect_transform)
-			luminosity_effect()
+		if (!testing)
+			if(first_step)
+				muzzle_effect(effect_transform)
+				first_step = FALSE
+			else if(!bumped)
+				tracer_effect(effect_transform)
+				luminosity_effect()
 
 		if(!hitscan)
 			sleep(step_delay)	//add delay between movement iterations if it's not a hitscan weapon
@@ -963,80 +979,13 @@
 			if(istype(M))
 				return 1
 
-// projectile/Test has been proven to be inaccurate and doesn't use the logic behind firing a real projectile.
-// impacttest uses the exact same logic, and thus, is perfectly accurate. It operates on the basis that whatever we hit will end the
-// flight of the projectile, thus, showing if we can hit a target or not.
+/proc/check_trajectory_raytrace(atom/movable/target, atom/movable/firer, var/proj, var/proc_path = null)
+	if (proc_path)
+		var/obj/item/projectile/trace = new proj(get_turf(firer))
 
-/obj/item/projectile/test/impacttest
-	hitscan = TRUE
+		firer.RegisterSignal(trace, COMSIG_TRACE_IMPACT, proc_path)
 
-/obj/item/projectile/test/impacttest/launch(atom/target, angle_offset, x_offset, y_offset)
-
-	var/turf/curloc = get_turf(src)
-	var/turf/targloc = get_turf(target)
-
-	if (!istype(targloc) || !istype(curloc))
-		return TRUE
-
-	if(targloc == curloc) //Shooting something in the same turf
-		on_impact(target)
-		qdel(src)
-		return FALSE
-
-	original = target
-
-	setup_trajectory(curloc, targloc, x_offset, y_offset, angle_offset) //plot the initial trajectory
-	Process()
-
-/obj/item/projectile/test/impacttest/Process()
-	spawn while(src && src.loc)
-		if(kill_count-- < 1)
-			on_impact(src.loc) //for any final impact behaviours
-			qdel(src)
-			return
-		if((!( current ) || loc == current))
-			current = locate(min(max(x + xo, 1), world.maxx), min(max(y + yo, 1), world.maxy), z)
-		if((x == 1 || x == world.maxx || y == 1 || y == world.maxy))
-			qdel(src)
-			return
-
-		trajectory.increment()	// increment the current location
-		location = trajectory.return_location(location)		// update the locally stored location data
-
-		if(!location)
-			qdel(src)	// if it's left the world... kill it
-			return
-
-		before_move()
-		Move(location.return_turf())
-		pixel_x = location.pixel_x
-		pixel_y = location.pixel_y
-
-		if(!bumped && !isturf(original))
-			if(loc == get_turf(original))
-				if(Bump(original))
-					return
-
-/obj/item/projectile/test/impacttest/on_impact(atom/A)
-
-	SEND_SIGNAL(src, COMSIG_TRACE_IMPACT, src, A)
-
-/obj/item/projectile/test/impacttest/Bump(atom/A, forced)
-
-	if(A == src)
-		return FALSE
-	if(A == firer)
-		loc = A.loc
-		return FALSE //go fuck yourself in another place pls
-
-	if(istype(A, /obj/item/projectile))
-		return FALSE
-
-	//stop flying
-	on_impact(A)
-
-	qdel(src)
-	return TRUE
+		trace.launch(target)
 
 //Helper proc to check if you can hit them or not.
 /proc/check_trajectory(atom/target as mob|obj, atom/firer as mob|obj, var/pass_flags=PASSTABLE|PASSGLASS|PASSGRILLE, flags=null)
